Rename Describe macro to LegacyDescribe
